event libs: default to building as dynamically loaded plugins
Event lib support as it has been isn't scaling well, at the low level
libevent and libev headers have a namespace conflict so they can't
both be built into the same image, and at the distro level, binding
all the event libs to libwebsockets.so makes a bloaty situation for
packaging, lws will drag in all the event libs every time.

This patch implements the plan discussed here

https://github.com/warmcat/libwebsockets/issues/1980

and refactors the event lib support so they are built into isolated
plugins and bound at runtime according to what the application says
it wants to use.  The event lib plugins can be packaged individually
so that only the needed sets of support are installed (perhaps none
of them if the user code is OK with the default poll() loop).  And
dependent user code can mark the specific event loop plugin package
as required so pieces are added as needed.

The eventlib-foreign example is also refactored to build the selected
lib support isolated.

A readme is added detailing the changes and how to use them.

https://libwebsockets.org/git/libwebsockets/tree/READMEs/README.event-libs.md

+++ minimal-examples/api-tests/api-test-lws_struct-json/main.c
@@ -780,7 +780,7 @@ done:
 
 		lws_struct_json_init_parse(&ctx, NULL, &a);
 
-		m = lejp_parse(&ctx, (uint8_t *)jig_conf, strlen(jig_conf));
+		m = lejp_parse(&ctx, (uint8_t *)jig_conf, (int)strlen(jig_conf));
 
 		if (m < 0 || !a.dest) {
 			lwsl_err("%s: line %d: JSON decode failed '%s'\n",
